requirements.txt
numpy
matplotlib
Pillow
jupyter
scikit-learn
